Replacing NDI Advanced SDK with NDI Standard SDK to comply with license
Processing.NDI.Advanced.h clearly states in its header:
"// This file is part of the NDI Advanced SDK and may not be distributed."
The title of `NDI Advanced License Agreement.pdf` reads "NewTek’s NDI® Embedded Software Development Kit (SDK) License Agreement".
`Processing.NDI.Lib.Licenses.txt` has a link to http://new.tk/ndisdk_embedded_license/ which is the identical text as `NDI Advanced License Agreement.pdf`.
So it is beyond reasonable to think that `NDI Advanced` and `NDI Embedded` are the same thing.
I diffed the text of `NDI SDK License Agreement.pdf` with `NDI Advanced License Agreement.pdf` and there are two differences:
1. `Advanced` "2. License:" appends "A separate license agreement with NewTek is required in order to commercially exploit or otherwise distribute any products that use or embed any NewTek Software, NDI® Software, or use part or all of the SDK and/or specific SDK, in any programmable chip or similar embedding device or technology (“NDI® Embedded Products”)."
2. `Advanced` fixes a few trivial punctuation and grammar typos.
  Examples:
  "compatability" -> "compatibility"
  "the the" -> "the"

That #1 "A separate license agreement with NewTek is required..." is the confusing part. So, **this** license agreement says that **another** license agreement is needed...but it does not explain what that other license agreement is or even how to obtain it.